Divide library into submodules
